src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/reinsert-insertion-point-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: inside.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Should be visible
